+++ src/db/migrations/0008_smooth_ender.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `downloads` ADD `source_id` text;--> statement-breakpoint
+ALTER TABLE `downloads` ADD `source_key` text;